requirements.txt
langgraph
langchain-openai
langchain-community
pydantic
python-dotenv
google-genai
streamlit
pandas
pypdf
wikipedia
duckduckgo-search
requests
PyGithub
